Move Environment field validation into container_validation.rs
Step 2 of splitting containers.rs. validate() (~270 lines: PHP
tuning, cron, backup schedule, Redis/Elasticsearch/MinIO/RabbitMQ,
Node, restart policy, WordPress admin fields) was private and only
called by save()/operate_inner() — no external re-export needed, just
import validate alongside the identity/platform and services/database
checks container_validation.rs already had. Every existing call site
(production and the ~20 tests that assert on it, several of which
also exercise compose/php_dockerfile in the same test) keeps working
unchanged via the updated `use crate::container_validation::{...}`
import. 3157 -> 2882 lines.
